updating tests for variable coeff of res
